insurance-rag-agent/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "insurance-rag-agent"
version = "0.1.0"
description = "Insurance RAG Agent for document analysis and question answering"
authors = [{name = "Your Name", email = "your.email@example.com"}]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
    "llama-index>=0.12.0",
    "llama-index-core>=0.12.0",
    "llama-index-llms-openai>=0.3.0",
    "llama-index-embeddings-openai>=0.3.0",
    "llama-index-readers-file>=0.3.0",
    "llama-index-retrievers-bm25>=0.6.0",
    "llama-index-vector-stores-chroma>=0.5.0",
    "chromadb>=0.5.0",
    "pdfplumber>=0.11.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "typer>=0.12.0",
    "rich>=13.0.0",
    "tiktoken>=0.7.0",
    "ragas>=0.1.0",
    "datasets>=2.18.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0"
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.black]
line-length = 88
target-version = ['py313']

[tool.mypy]
python_version = "3.13"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
